requirements.txt
requests==2.31.0
python-dotenv==1.0.0
supabase==2.11.0
openai==1.12.0
google-generativeai==0.3.2
beautifulsoup4==4.12.3
pandas==2.2.0
curl-cffi==0.5.10
selenium
webdriver-manager
